Integration tests

See merge request code-forcer-ise-5/code_forcer!67

+++ backend/src/features/students/delete_student.py
@@ -1,4 +1,5 @@
-from fastapi import status, APIRouter
+from fastapi import status, APIRouter, HTTPException
+from pydantic import EmailStr
 
 from src.container import container
 from .interfaces import IStudentsRepository
@@ -7,7 +8,7 @@ router = APIRouter()
 
 
 @router.delete("/students/{email}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_student(email: str) -> None:
+async def delete_student(email: EmailStr) -> None:
     DeleteStudentCommandHandler(
         container[IStudentsRepository]
     ).handle(email)
@@ -17,5 +18,11 @@ class DeleteStudentCommandHandler:
     def __init__(self, students_repository: IStudentsRepository):
         self.students_repository = students_repository
 
-    def handle(self, email: str) -> None:
+    def handle(self, email: EmailStr) -> None:
+        if not self.students_repository.email_exists(email):
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail='Student with given email is not found'
+            )
+
         self.students_repository.delete_student(email)
